requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
langchain-text-splitters
sentence_transformers
langchain_huggingface
langchain_chroma
groq
langchain_groq
pymupdf
langsmith
streamlit
fastapi
uvicorn
wikipedia
arxiv
faiss-cpu
pdfplumber